smd: create quirk for non standard behaviour
It has been observed on msm899(2/4) that the bootloaders / LK leave the link
between the RPM and APPs processor in an undesirable state.  When we boot into
Linux we expect the remote channel to be in the SMD_CHANNEL_CLOSED state
unfortunately this is not the case on the msm899(2/4).  Instead the RX channel
is sitting in a state of SMD_CHANNEL_CLOSING , in order to deal with this
anomaly and to deal with any future potential issues in which we need to work
around unexpected behaviour.   Create a quirks DT entry which will run any
workarounds to allow proper operation.

In this case we are able to get the remote (RX) back into the desired initial
starting state by sequencing the states on our end (host / APPs processor).
If we try forcing the RX end directly into the desired state it will be exhibit
erratic behaviour.   Therefor we _MUST_ follow the correct sequence of
states: 0, 1, 2....

Signed-off-by: Jeremy McNicoll <jeremymc@redhat.com>